Refactor : rebase 1차 분리 및 styled component 적용
