+++ src/Pages/AdminPage/ModalAddVacationAll.jsx
@@ -1,8 +1,16 @@
 import { useState } from 'react';
+
+import styled from 'styled-components';
+
 import Button from '@mui/material/Button';
 import Slider from '@mui/material/Slider';
 
-const AddVacation = ({ setIsOpen, attendUser, addVacation, minusVacation }) => {
+const ModalAddVacationAll = ({
+  setIsOpen,
+  attendUser,
+  addVacation,
+  minusVacation,
+}) => {
   const [value, setValue] = useState(0);
   // console.log('attend', attendUser);
 
@@ -29,7 +37,7 @@ const AddVacation = ({ setIsOpen, attendUser, addVacation, minusVacation }) => {
     setValue(event.target.value);
   };
   return (
-    <div className="modal">
+    <ModalAddVacationAllBody>
       <h1>일괄 휴가 변경</h1>
       <h3>참가 중인 모든 사용자들의 휴가를 증가하거나 감소합니다.</h3>
       <h3>얼마나 변경하시겠습니까?</h3>
@@ -46,8 +54,17 @@ const AddVacation = ({ setIsOpen, attendUser, addVacation, minusVacation }) => {
       />
       <Button onClick={() => handleCloseModal(true)}>확인</Button>
       <Button onClick={() => handleCloseModal(false)}>취소</Button>
-    </div>
+    </ModalAddVacationAllBody>
   );
 };
 
-export default AddVacation;
+const ModalAddVacationAllBody = styled.div`
+  position: absolute;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  background-color: rgba(216, 216, 216, 0.9);
+`;
+
+export default ModalAddVacationAll;
